+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### 0.1.0 [2017-05-22]
 - Versão inicial
+- Added Configuração na BaseEntity
 
 ## Informações adicionais
 - Todas as mudanças notáveis neste projeto serão documentadas neste arquivo.